Debugging _transform_expression
